backend/requirements.txt
# ============================================
# DDCPTS Backend Dependencies
# Python 3.11+
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0  # For async PostgreSQL support

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6

# SMS Providers
africastalking==1.2.6
twilio==8.11.0

# M-Pesa / Payment
requests==2.31.0
httpx==0.26.0

# Background Tasks
celery[redis]==5.3.6
redis==5.0.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
phonenumbers==8.13.27
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For TestClient
faker==22.0.0  # For generating test data
freezegun==1.4.0  # For time mocking

# Code Quality
ruff==0.1.14
mypy==1.8.0
black==24.1.1

# Development
ipython==8.20.0
watchfiles==0.21.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
